oct 26, 2023 update

+++ ICPC/2022 ICPC East Central/2022ICPCEastCentral_D.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 typedef long long ll;
-typedef long long ld;
+typedef long double ld;
 
 
 int main() {
